Spacing between chracters are now working as intended. To : glyphData.myAdvanceX = slot->metrics.width >> 6; From : glyphData.myAdvanceX = slot->advance.x;
Space between words now works as intended on the bold font in use.
Further testing is needed on more fonts to confirm that it works as intended on more than one.
Using the width of minor char 'x' to get the width and then I divide it by 256.
fontData->myWordSpacing = space->metrics.width >> 8;