Added filepath to system fonts if nothing is inputted more than a font name.
If you load a font with arial.ttf as example

text = new CText("Arial.ttf", pxSize, borderSize);
Then it will go to C:\\Windows\\Fonts\\
and get the Arial.ttf from there.

Fonts that is not in the system by default will have to be loaded as usual with the entire filepath.


Former-commit-id: 31d96d98455ac9c0d5f3e10ed05fd393f67ee841