Optional sampler read from shader files
Rewrote the sampler part of shaders to be optional, if no sampler is set the previous one in use will be used.
Might have to make a SettingsManager to handle the different samplers. Either that or I'll just use a base sampler and then throw in a bunch of other samplers behind the scene.


Former-commit-id: 9f41a4e5aa158285d7347d03debcb95987e41cd2