Adding RenderContext calls
Working on adding RenderContext calls to everything, slowly making decent progress. About to start replacing Quad Rendering wherever I do it with the generic quad class that contains very little, might have to rewrite it a bit more if I want to be able to use any shader with it.


Former-commit-id: 4b8b470c06c487bc896e9d7a0a1905d4071734bd