Changed the model import
the model import now creates the different buffers for the model, including the input layouts for rendering.
This means that all models creates the instance wrapper objects even if they don't end up using them.
This might be changed in the future, but it becomes easier to test performance this way. Most things will be instanced in the future tho, it will be individual objects that will not be instaned, if the count of models is less than 10 or so then we should probably not instance it.


Former-commit-id: 8595627b520d4ca5b27b4022a597ed971b7f1389