Resolved driver crash & Viewport now set
The default viewport now gets set at creation. Still need a way of resetting it after shadowpass and post processing.

The driver crash was caused by using the "instance count" from the instance wrapper, this number could be either 0 or 5000, none of these values were correct, so I wrote a new function that could get the size of the orientation array and then used that in a new DrawIndexedInstanced function for Model* and not BaseModel*


Former-commit-id: aa73fcb95a23ea8b6662e35d1113238527f3d06d